New structure for the whole repository: de\hsos\prog3\ followed by the actual tasks. User repos are stored in the src folder under tasks, thus using the same ressources (i.e. on lib).
Changes can now be made "from the bottom up", first merging different user repos to one macro repo und finally the master.
