--- ab01/ab01_2/src/de/hsos/prog3/dieteri/ab01/Main.java
@@ -1,17 +1,0 @@
-package de.hsos.prog3.dieteri.ab01;
-
-import de.hsos.prog3.audio.SimpleAudioPlayer;
-
-import java.io.IOException;
-import java.net.URL;
-
-public class Main {
-
-    public static void main(String[] args) throws IOException {
-	    URL url = Main.class.getResource("/Baritone.wav");
-        SimpleAudioPlayer player = new SimpleAudioPlayer(url);
-        player.setDebug(false);
-        player.verboseLogging(true);
-        player.play(0);
-    }
-}
