+++ de/hsos/prog3/ab01/ab01_2/src/dieteri/ab01/Main.java
@@ -1,0 +1,17 @@
+package src.dieteri.ab01;
+
+import de.hsos.prog3.audio.SimpleAudioPlayer;
+
+import java.io.IOException;
+import java.net.URL;
+
+public class Main {
+
+    public static void main(String[] args) throws IOException {
+	    URL url = Main.class.getResource("/res/Drum.wav");
+        SimpleAudioPlayer player = new SimpleAudioPlayer(url);
+        player.setDebug(false);
+        player.verboseLogging(true);
+        player.play(0);
+    }
+}
